test-doc-generator/requirements.txt
openpyxl>=3.1.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
